+++ app/helpers/application_helper.rb
@@ -80,7 +80,7 @@ module ApplicationHelper
     checked ? icon("fas", "check-square") : icon("far", "square")
   end
 
-  def simple_markdown(text, table_class = "")
+  def simple_markdown_old(text, table_class = "")
     markdown = Redcarpet::Markdown.new(Redcarpet::Render::HTML, no_intra_emphasis: true, fenced_code_blocks: true, autolink: true, strikethrough: true, superscript: true, tables: true, lax_spacing: true, space_after_headers: true, underline: true, highlight: true, footnotes: true)
     result = replace_numbers_with_ascii(text.to_s)
     result = markdown.render(result)