+++ db/migrate/20190320170559_add_cache_busters_to_designs.rb
@@ -1,0 +1,6 @@
+class AddCacheBustersToDesigns < ActiveRecord::Migration[6.0]
+  def change
+    add_column :designs, :pdf_cache_busted_at, :datetime
+    add_column :designs, :coverage_cache_busted_at, :datetime
+  end
+end